refactor CMakeLists and directory structure
