+++ include/gurkenlaeufer/Variant.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <string>
+#include <cstdint>
 
 class Variant {
 public:
